refactor: split aggregate transform implementation
